+++ src/component/Editor/InitialEditor.js
@@ -1,6 +1,6 @@
 export default function InitialEditor({ $target }) {
   const $image = document.createElement('img');
-  $image.setAttribute('src', '/src/img/DongTion.png');
+  $image.setAttribute('src', '/src/assets/img/DongTion.png');
   $image.className = 'main-img';
 
   this.removeMainEditor = () => {
