+++ src/main.js
@@ -1,5 +1,5 @@
 import App from './App.js';
-import { $ } from './lib/utils.js';
+import { $ } from './lib/dom.js';
 const $app = $('#app');
 
 new App({ $target: $app });
